+++ release_notes.md
@@ -1,5 +1,9 @@
 # v0.5.1
 
+### New Features
+
+- Added version to window title bar
+
 ### Fixes
 
 - Showing online players on server